Move the position helper to global helpers and rename variables and params
