+++ packages/vue3/src/_dev/App.vue
@@ -1,28 +1,25 @@
 <script setup lang="ts">
 import 'uno.css';
 import '../assets/main.css';
-import { Badge } from '~/components';
-import type { BadgeState } from '~/types';
-const badgeLabel = 'Badge';
-const labelStates: BadgeState[] = ['danger', 'info', 'neutral', 'success', 'warning'];
+import { SecondaryButton, Tooltip } from '~/components';
 </script>
 
 <template>
   <div class="container">
-    <Badge v-for="state in labelStates" :key="state" :state="state" :size="24">
-      {{ badgeLabel }}
-    </Badge>
+    <Tooltip label="Add to favorites" position="top">
+      <SecondaryButton>
+        Hover
+      </SecondaryButton>
+    </Tooltip>
   </div>
 </template>
 
 <style scoped>
 .container {
   display: flex;
-  gap: 10px;
   align-items: center;
   justify-content: center;
-  width: 40vw;
-  height: 40vh;
-  margin: 10%;
+  width: 100vw;
+  height: 100vh;
 }
 </style>
